+++ src/bakery_manager_mobile/lib/pages/add_recipe_page.dart
@@ -1,0 +1,103 @@
+import 'package:flutter/material.dart';
+import '../services/api_service.dart';
+
+class AddRecipePage extends StatefulWidget {
+  const AddRecipePage({super.key});
+
+  @override
+  State<AddRecipePage> createState() => _AddRecipePageState();
+}
+
+class _AddRecipePageState extends State<AddRecipePage> {
+  final TextEditingController recipeNameController = TextEditingController();
+  final TextEditingController ingredientsController = TextEditingController();
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Add Recipe', style: TextStyle(color: Colors.white)),
+        backgroundColor: Colors.orange,
+        leading: IconButton(
+          icon: const Icon(Icons.arrow_back, color: Colors.white),
+          onPressed: () {
+            Navigator.pop(context);
+          },
+        ),
+      ),
+      body: Padding(
+        padding: const EdgeInsets.all(16.0),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.stretch,
+          children: [
+            const Text(
+              'Add a New Recipe',
+              style: TextStyle(fontSize: 24, fontWeight: FontWeight.bold),
+              textAlign: TextAlign.center,
+            ),
+            const SizedBox(height: 16),
+            TextField(
+              controller: recipeNameController,
+              decoration: const InputDecoration(
+                hintText: 'Recipe Name',
+                border: OutlineInputBorder(
+                  borderRadius: BorderRadius.all(Radius.circular(10)),
+                ),
+              ),
+            ),
+            const SizedBox(height: 16),
+            TextField(
+              controller: ingredientsController,
+              keyboardType: TextInputType.multiline,
+              maxLines: null,
+              decoration: InputDecoration(
+                hintText: 'Recipe Ingredients',
+                border: OutlineInputBorder(
+                  borderRadius: BorderRadius.circular(10.0),
+                ),
+              ),
+            ),
+            const SizedBox(height: 16),
+            ElevatedButton(
+              style: ElevatedButton.styleFrom(
+                backgroundColor: Colors.orange,
+                padding: const EdgeInsets.symmetric(vertical: 16),
+                shape: RoundedRectangleBorder(
+                  borderRadius: BorderRadius.circular(10),
+                ),
+              ),
+              onPressed: () async {
+                String recipeName = recipeNameController.text;
+                String ingredients = ingredientsController.text;
+
+                // Save recipe
+                Map<String, dynamic> response = await ApiService.addRecipe(
+                    recipeName: recipeName, ingredients: ingredients);
+                if (response['status'] == 'success') {
+                  if (context.mounted) {
+                    ScaffoldMessenger.of(context).showSnackBar(const SnackBar(
+                        content: Text('Recipe added successfully')));
+                    Navigator.pop(context);
+                  }
+                } else {
+                  if (context.mounted) {
+                    ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+                        content: Text(
+                            'Failed to add recipe: ${response['reason']}')));
+                  }
+                }
+              },
+              child: const Text('Save Recipe'),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  @override
+  void dispose() {
+    recipeNameController.dispose();
+    ingredientsController.dispose();
+    super.dispose();
+  }
+}