+++ CHANGELOG.md
@@ -15,11 +15,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Static Home Page 
 - Created a login page.
 - Added custom fonts.
+- Added api service to handle api calls.
+- Added api calls for login, register, get recipes, and add recipe.
+- Added navigator observer to handle to reload when navigating to a back to a page.
 
 ### Changed
 
 - Moved Most Code Outside main.dart Into Individual Files Based off Page.
 - Updated Changelog with resources for ease of access
+- Moved add recipe page to its own file.
 
 ## [0.0.2] - 2024-08-28